pyproject.toml
[project]
name = "nemotron-speech"
version = "0.1.0"
description = "Nemotron-Speech-ASR streaming STT service for Pipecat on DGX Spark"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["asr", "speech-to-text", "streaming", "nemo", "pipecat", "nvidia"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "gdown>=5.0.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "numpy>=1.26.0",
    "loguru>=0.7.0",
    # Pipecat bot dependencies
    "pipecat-ai[cartesia,daily,google,local-smart-turn-v3,openai,runner,silero,webrtc]==0.0.98",
    # Local Riva TTS
    "nvidia-riva-client>=2.16.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
# NeMo ASR dependencies - install separately due to complex CUDA requirements
nemo = [
    "nemo_toolkit[asr]>=2.6.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
modal = [
    "modal>=1.3.0.post1",
]
bot = [
   "pipecat-ai-cli>=0.1.8",
]

[project.urls]
Homepage = "https://github.com/yourusername/nemotron-speech"
Documentation = "https://github.com/yourusername/nemotron-speech#readme"
Repository = "https://github.com/yourusername/nemotron-speech"

[project.scripts]
nemotron-speech = "nemotron_speech.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nemotron_speech"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "aiortc>=1.14.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
